fix(plex link): Improve mobile deep linking
The previous method of generating Plex deep links was unreliable on mobile devices. This change updates the logic to use the `plex://preplay` command, which is specifically designed for navigating to item details within the native Plex app. This ensures a more robust and consistent deep linking experience on mobile.

Also, includes a minor update to the build number.